% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{balanced_accuracy}
\alias{balanced_accuracy}
\title{Balanced Accuracy}
\usage{
balanced_accuracy(y, y_hat, sample_weight = NULL, adjusted = FALSE)
}
\arguments{
\item{y, y_hat}{real values and fitted values.}

\item{sample_weight}{numeric vector of sample weights, default \code{NULL}.}

\item{adjusted}{if \code{TRUE}, adjust the score for chance so that random
performance scores 0 and perfect performance scores 1.}
}
\value{
balanced accuracy score, a numeric scalar.
}
\description{
Compute balanced accuracy for classification, which is the average of
recall obtained on each class. This implementation aligns with
\code{sklearn.metrics.balanced_accuracy_score}.
}
\author{
Zhang Jiaqi
}
