% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bals_svm.R
\name{bals_svm}
\alias{bals_svm}
\title{Bounded Asymmetric Least Squares Loss Support Vector Machine}
\usage{
bals_svm(
  X,
  y,
  C = 1,
  kernel = c("linear", "rbf", "poly"),
  gamma = 1/ncol(X),
  degree = 3,
  coef0 = 0,
  p = 0.5,
  lambda = 1,
  eps = 1e-05,
  eps.hq = 0.01,
  max.steps = 4000,
  hq.steps = 10,
  fit_intercept = TRUE,
  dual_optimizer = clip_dcd_optimizer,
  dual_optimizer_option = NULL
)
}
\arguments{
\item{X, y}{dataset and label.}

\item{C}{plenty term.}

\item{kernel}{kernel function. The definitions of various kernel functions are as follows:
\describe{
    \item{linear:}{\eqn{u'v}{u'*v}}
    \item{poly:}{\eqn{(\gamma u'v + coef0)^{degree}}{(gamma*u'*v + coef0)^degree}}
    \item{rbf:}{\eqn{e^{(-\gamma |u-v|^2)}}{exp(-gamma*|u-v|^2)}}
}}

\item{gamma}{parameter for \code{'rbf'} and \code{'poly'} kernel. Default \code{gamma = 1/ncol(X)}.}

\item{degree}{parameter for polynomial kernel, default: \code{degree = 3}.}

\item{coef0}{parameter for polynomial kernel,  default: \code{coef0 = 0}.}

\item{p}{parameter for bounded asymmetric least squares loss.}

\item{lambda}{parameter for bq loss (loss increase speed).}

\item{eps}{the precision of the optimization algorithm.}

\item{eps.hq}{the precision of the optimization algorithm.}

\item{max.steps}{the number of iterations to solve the optimization problem.}

\item{hq.steps}{the number of iterations of HQ}

\item{fit_intercept}{if set \code{fit_intercept = TRUE},
the function will evaluates intercept.}

\item{dual_optimizer}{default optimizer is \code{clip_dcd_optimizer}.}

\item{dual_optimizer_option}{optimizer options.}
}
\value{
return \code{SVMClassifier} object.
}
\description{
\code{bals_svm} is an R implementation of BALS-SVM
}
\author{
Zhang Jiaqi.
}
