% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizers.R
\name{batch_iter}
\alias{batch_iter}
\title{A Batch Indices Iterator}
\usage{
batch_iter(n, batch_size)
}
\arguments{
\item{n}{number of samples.}

\item{batch_size}{batch size (0 < batch_size <= n).}
}
\description{
A Batch Indices Iterator
}
