% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{binaryf1score}
\alias{binaryf1score}
\title{Binary F1 Score
return F1 Score for two vector}
\usage{
binaryf1score(y, y_hat, positive = c("Class-1", "Class-2"))
}
\arguments{
\item{y, y_hat}{real values and fitted values.}

\item{positive}{positive class, default "Class-1".}
}
\description{
Binary F1 Score
return F1 Score for two vector
}
\author{
Zhang Jiaqi
}
