% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bq_svm.R
\name{bq_svm}
\alias{bq_svm}
\title{Bounded Quantile Loss Support Vector Machine}
\usage{
bq_svm(
  X,
  y,
  C = 1,
  kernel = c("linear", "rbf", "poly"),
  gamma = 1/ncol(X),
  degree = 3,
  coef0 = 0,
  tau = 1,
  lambda = 1,
  eps = 1e-05,
  eps.cccp = 0.01,
  max.steps = 4000,
  cccp.steps = 10,
  batch_size = nrow(X)/10,
  solver = c("dual", "primal"),
  fit_intercept = TRUE,
  optimizer = pegasos,
  reduce_set = NULL,
  dual_optimizer = clip_dcd_optimizer,
  dual_optimizer_option = NULL,
  ...
)
}
\arguments{
\item{X, y}{dataset and label.}

\item{C}{plenty term.}

\item{kernel}{kernel function. The definitions of various kernel functions are as follows:
\describe{
    \item{linear:}{\eqn{u'v}{u'*v}}
    \item{poly:}{\eqn{(\gamma u'v + coef0)^{degree}}{(gamma*u'*v + coef0)^degree}}
    \item{rbf:}{\eqn{e^{(-\gamma |u-v|^2)}}{exp(-gamma*|u-v|^2)}}
}}

\item{gamma}{parameter for \code{'rbf'} and \code{'poly'} kernel. Default \code{gamma = 1/ncol(X)}.}

\item{degree}{parameter for polynomial kernel, default: \code{degree = 3}.}

\item{coef0}{parameter for polynomial kernel,  default: \code{coef0 = 0}.}

\item{tau}{parameter for bq loss.}

\item{lambda}{parameter for bq loss (loss increase speed).}

\item{eps}{the precision of the optimization algorithm.}

\item{eps.cccp}{the precision of the optimization algorithm.}

\item{max.steps}{the number of iterations to solve the optimization problem.}

\item{cccp.steps}{the number of iterations of CCCP.}

\item{batch_size}{mini-batch size for primal solver.}

\item{solver}{\code{"dual"} and \code{"primal"} are available.}

\item{fit_intercept}{if set \code{fit_intercept = TRUE},
the function will evaluates intercept.}

\item{optimizer}{default primal optimizer pegasos.}

\item{reduce_set}{reduce set for reduce SVM, default \code{reduce_set = NULL}.}

\item{dual_optimizer}{default optimizer is \code{clip_dcd_optimizer}.}

\item{dual_optimizer_option}{optimizer options.}

\item{...}{unused parameters.}
}
\value{
return \code{SVMClassifier} object.
}
\description{
\code{bq_svm} is an R implementation of bq-SVM
}
\author{
Zhang Jiaqi.
}
