% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_systems_solver.R
\name{cholsolve}
\alias{cholsolve}
\title{Solve Linear Systems Use Cholesky Decomposition.}
\usage{
cholsolve(a, b)
}
\arguments{
\item{a}{a symmetric square matrix containing the coefficients of the linear system.}

\item{b}{a vector or matrix giving the right-hand side(s).}
}
\description{
Solve Linear Systems Use Cholesky Decomposition.
}
\author{
Zhang Jiaqi.
}
