% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cL2p_ls_tsvm.R
\name{cL2p_ls_tsvm}
\alias{cL2p_ls_tsvm}
\title{Capped L2,p Least Squares Twin Support Vector Machine}
\usage{
cL2p_ls_tsvm(
  X,
  y,
  C1 = 1,
  C2 = C1,
  kernel = c("linear", "rbf", "poly"),
  gamma = 1/ncol(X),
  degree = 3,
  coef0 = 0,
  p = 1,
  epsilon1 = 1,
  epsilon2 = epsilon1,
  eps.irls = 0.01,
  irls.steps = 10,
  fit_intercept = TRUE,
  reduce_set = 1
)
}
\arguments{
\item{X, y}{dataset and label.}

\item{C1, C2}{plenty term.}

\item{kernel}{kernel function. The definitions of various kernel functions are as follows:
\describe{
    \item{linear:}{\eqn{u'v}{u'*v}}
    \item{poly:}{\eqn{(\gamma u'v + coef0)^{degree}}{(gamma*u'*v + coef0)^degree}}
    \item{rbf:}{\eqn{e^{(-\gamma |u-v|^2)}}{exp(-gamma*|u-v|^2)}}
}}

\item{gamma}{parameter for \code{'rbf'} and \code{'poly'} kernel. Default \code{gamma = 1/ncol(X)}.}

\item{degree}{parameter for polynomial kernel, default: \code{degree = 3}.}

\item{coef0}{parameter for polynomial kernel,  default: \code{coef0 = 0}.}

\item{p}{parameter for L2,p norm.}

\item{epsilon1, epsilon2}{truncation parameter.}

\item{eps.irls}{the precision of the optimization algorithm.}

\item{irls.steps}{the number of iterations to solve the optimization problem.}

\item{fit_intercept}{if set \code{fit_intercept = TRUE},
the function will evaluates intercept.}

\item{reduce_set}{reduce set for reduce SVM, default \code{reduce_set = NULL}.}
}
\value{
return \code{TSVMClassifier} object.
}
\description{
\code{cL2p_ls_tsvm} is an R implementation of CL2,p-LS-TSVM
}
\author{
Zhang Jiaqi.
}
