% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clip_dcd_optimizer.R
\name{clip_dcd_optimizer}
\alias{clip_dcd_optimizer}
\title{Clipping Dual Coordinate Decent Optimizer}
\usage{
clip_dcd_optimizer(
  H,
  q,
  lb,
  ub,
  eps = 1e-05,
  max.steps = 200,
  u = (lb + ub)/2,
  ...
)
}
\arguments{
\item{H, q}{input matrices.}

\item{lb, ub}{lower bound and upper bound.}

\item{eps, max.steps}{error and maximum iterations.}

\item{u}{init solution.}

\item{...}{redundant parameters.}
}
\value{
return results list
}
\description{
Clipping Dual Coordinate Decent Optimizer
}
\author{
Zhang Jiaqi.
}
