% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closs_tbsvm.R
\name{closs_tbsvm}
\alias{closs_tbsvm}
\title{C-loss Least Squares Twin Bounded Support Vector Machine}
\usage{
closs_tbsvm(
  X,
  y,
  C1 = 1,
  C2 = C1,
  C3 = 1e-07,
  C4 = C3,
  kernel = c("linear", "rbf", "poly"),
  gamma = 1/ncol(X),
  degree = 3,
  coef0 = 0,
  sigma1 = 1,
  sigma2 = sigma1,
  eps.hq = 0.01,
  hq.steps = 10,
  fit_intercept = TRUE,
  reduce_set = NULL
)
}
\arguments{
\item{X, y}{dataset and label.}

\item{C1, C2, C3, C4}{plenty term.}

\item{kernel}{kernel function. The definitions of various kernel functions are as follows:
\describe{
    \item{linear:}{\eqn{u'v}{u'*v}}
    \item{poly:}{\eqn{(\gamma u'v + coef0)^{degree}}{(gamma*u'*v + coef0)^degree}}
    \item{rbf:}{\eqn{e^{(-\gamma |u-v|^2)}}{exp(-gamma*|u-v|^2)}}
}}

\item{gamma}{parameter for \code{'rbf'} and \code{'poly'} kernel. Default \code{gamma = 1/ncol(X)}.}

\item{degree}{parameter for polynomial kernel, default: \code{degree = 3}.}

\item{coef0}{parameter for polynomial kernel,  default: \code{coef0 = 0}.}

\item{sigma1, sigma2}{parameters for C-loss function.}

\item{eps.hq}{the precision of the optimization algorithm.}

\item{hq.steps}{the number of iterations to solve the optimization problem.}

\item{fit_intercept}{if set \code{fit_intercept = TRUE},
the function will evaluates intercept.}

\item{reduce_set}{reduce set for reduce SVM, default \code{reduce_set = NULL}.}
}
\value{
return \code{TSVMClassifier} object.
}
\description{
\code{closs_tbsvm} is an R implementation of CL2,p-LS-TSVM
}
\author{
Zhang Jiaqi.
}
