% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hinge_nhsvm.R
\name{coef.NHSVMClassifier}
\alias{coef.NHSVMClassifier}
\title{Coef Method for Nonparallel Hyperplane Support Vector Machine}
\usage{
\method{coef}{NHSVMClassifier}(object, ...)
}
\arguments{
\item{object}{a fitted object of class inheriting from \code{NHSVMClassifier}.}

\item{...}{unused parameter.}
}
\description{
Coef Method for Nonparallel Hyperplane Support Vector Machine
}
\author{
Zhang Jiaqi
}
