% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hinge_svm.R
\name{coef.SVMClassifier}
\alias{coef.SVMClassifier}
\title{Coef Method for Support Vector Machine}
\usage{
\method{coef}{SVMClassifier}(object, ...)
}
\arguments{
\item{object}{a fitted object of class inheriting from \code{SVMClassifier}.}

\item{...}{unused parameter.}
}
\description{
Coef Method for Support Vector Machine
}
\author{
Zhang Jiaqi
}
