% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eps_svr.R
\name{coef.SVMRegressor}
\alias{coef.SVMRegressor}
\title{Coef Method for Support Vector Regression}
\usage{
\method{coef}{SVMRegressor}(object, ...)
}
\arguments{
\item{object}{a fitted object of class inheriting from \code{SVMRegressor}.}

\item{...}{unused parameter.}
}
\description{
Coef Method for Support Vector Regression
}
\author{
Zhang Jiaqi
}
