% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hinge_tsvm.R
\name{coef.TSVMClassifier}
\alias{coef.TSVMClassifier}
\title{Coef Method for Twin Support Vector Machine}
\usage{
\method{coef}{TSVMClassifier}(object, ...)
}
\arguments{
\item{object}{a fitted object of class inheriting from \code{TSVMClassifier}.}

\item{...}{unused parameter.}
}
\description{
Coef Method for Twin Support Vector Machine
}
\author{
Zhang Jiaqi
}
