% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twinsvm.R
\name{coef.twinsvm}
\alias{coef.twinsvm}
\title{Coef Method for Twin Support Vector Machines}
\usage{
\method{coef}{twinsvm}(object, ...)
}
\arguments{
\item{object}{object of class `twinsvm`.}

\item{...}{unused parameter.}
}
\description{
This function solve coefficients based upon a model trained by \code{twinsvm}.
}
\examples{
library(manysvms)
data('iris')

m <- 100
X <- iris[0:m, 1:2]
y <- iris[0:m, 5]
model <- twinsvm(X, y, kernel = 'linear')
pred <-predict(model, X, y)
}
\author{
Zhang Jiaqi
}
