% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizers.R
\name{conjugate_gradient_method}
\alias{conjugate_gradient_method}
\title{Conjugate Gradient Method for Solving Linear Equation Ax = b}
\usage{
conjugate_gradient_method(A, b, x0, eps = 1e-05, max.steps = 2 * nrow(A))
}
\arguments{
\item{A, b}{matrix of linear equation Ax = b (Since A is a PSD matrix).}

\item{x0}{initial point.}

\item{eps}{the precision of the optimization algorithm.}

\item{max.steps}{the number of iterations to solve the optimization problem.}
}
\value{
return optimal solution.
}
\description{
Conjugate Gradient Method for Solving Linear Equation Ax = b
}
\author{
Zhang Jiaqi.
}
