% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbsvm.R
\name{cv.mbsvm}
\alias{cv.mbsvm}
\title{Computes K-fold Cross-Validation Accuracy for MBSVM}
\usage{
cv.mbsvm(
  X,
  y,
  K = 5,
  C = 1,
  kernel = c("linear", "rbf", "poly"),
  gamma = 1/ncol(X),
  degree = 3,
  coef0 = 0,
  reg = 1e-07,
  kernel_rect = 1,
  tol = 1e-05,
  max.steps = 200,
  rcpp = TRUE,
  shuffle = TRUE,
  seed = NULL,
  threads.num = parallel::detectCores() - 1
)
}
\arguments{
\item{X, y}{dataset and label.}

\item{K}{Number of folds.}

\item{C}{plenty term.}

\item{kernel}{kernel type. Default value \code{kernel = 'linear'}.}

\item{gamma}{parameter for \code{'rbf'} and \code{'poly'} kernel. Default \code{gamma = 1/ncol(X)}.}

\item{degree}{parameter for polynomial kernel, default: \code{degree = 3}.}

\item{coef0}{parameter for polynomial kernel,  default: \code{coef0 = 0}.}

\item{reg}{regularization term to take care of problems due to ill-conditioning in dual problem.}

\item{kernel_rect}{set kernel size. \code{0<= kernel_rect <= 1}}

\item{tol}{the precision of the optimization algorithm.}

\item{max.steps}{the number of iterations to solve the optimization problem.}

\item{rcpp}{speed up your code with Rcpp, default \code{rcpp = TRUE}.}

\item{shuffle}{if set \code{shuffle==TRUE}, This function will shuffle the dataset.}

\item{seed}{random seed for \code{shuffer} option.}

\item{threads.num}{The number of threads used for parallel execution.}
}
\description{
Computes K-fold Cross-Validation Accuracy for MBSVM
}
\author{
Zhang Jiaqi
}
