% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ramptwinKsvm.R
\name{cv.ramptwinKsvm}
\alias{cv.ramptwinKsvm}
\title{Computes K-fold Cross-Validation Accuracy for ramp Twin K svm}
\usage{
cv.ramptwinKsvm(
  X,
  y,
  K = 5,
  Ck = rep(1, 4),
  sk = rep(0.5, 4),
  kernel = c("linear", "rbf", "poly"),
  gamma = 1/ncol(X),
  degree = 3,
  coef0 = 0,
  reg = 1,
  kernel_rect = 1,
  eps = 0.1,
  tol = 1e-05,
  cccp.steps = 30,
  max.steps = 300,
  rcpp = TRUE,
  shuffer = TRUE,
  seed = NULL
)
}
\arguments{
\item{X}{a new data frame for predicting.}

\item{y}{a label data frame corresponding to X.}

\item{K}{number of folds.}

\item{Ck}{plenty term vector.}

\item{sk}{parameter for ramp loss.}

\item{kernel}{kernel function.}

\item{gamma}{rbf kernel parameter.}

\item{degree}{parameter for polynomial kernel, default: \code{degree = 3}.}

\item{coef0}{parameter for polynomial kernel,  default: \code{coef0 = 0}.}

\item{reg}{regularization term.}

\item{kernel_rect}{set kernel size. \code{0<= kernel_rect <= 1}}

\item{eps}{parameter for rest class.}

\item{tol}{the precision of the optimization algorithm.}

\item{cccp.steps}{the number of iterations of Concave–Convex Procedure (CCCP).}

\item{max.steps}{the number of iterations to solve the optimization problem.}

\item{rcpp}{speed up your code with Rcpp, default \code{rcpp = TRUE}.}

\item{shuffer}{if set \code{shuffer==TRUE}, This function will shuffle the dataset.}

\item{seed}{random seed for \code{shuffer} option.}
}
\description{
Computes K-fold Cross-Validation Accuracy for ramp Twin K svm
}
\author{
Zhang Jiaqi
}
