% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twinKsvm.R
\name{cv.twinKsvm}
\alias{cv.twinKsvm}
\title{Computes K-fold Cross-Validation Accuracy for Twin K svm}
\usage{
cv.twinKsvm(
  X,
  y,
  K = 5,
  C1 = 1,
  C3 = 1,
  kernel = c("linear", "rbf", "poly"),
  gamma = 1/ncol(X),
  degree = 3,
  coef0 = 0,
  reg = 1e-07,
  kernel_rect = 1,
  eps = 0.1,
  tol = 1e-05,
  max.steps = 200,
  rcpp = TRUE,
  shuffle = TRUE,
  seed = NULL,
  threads.num = parallel::detectCores() - 1
)
}
\arguments{
\item{X}{a new data frame for predicting.}

\item{y}{a label data frame corresponding to X.}

\item{K}{number of folds.}

\item{C1, C3}{plenty term vector.}

\item{kernel}{kernel function.}

\item{gamma}{rbf kernel parameter.}

\item{degree}{parameter for polynomial kernel, default: \code{degree = 3}.}

\item{coef0}{parameter for polynomial kernel,  default: \code{coef0 = 0}.}

\item{reg}{regularization term.}

\item{kernel_rect}{set kernel size. \code{0<= kernel_rect <= 1}}

\item{eps}{parameter for rest class.}

\item{tol}{the precision of the optimization algorithm.}

\item{max.steps}{the number of iterations to solve the optimization problem.}

\item{rcpp}{speed up your code with Rcpp, default \code{rcpp = TRUE}.}

\item{shuffle}{if set \code{shuffer==TRUE}, This function will shuffle the dataset.}

\item{seed}{random seed for \code{shuffle} option.}

\item{threads.num}{The number of threads used for parallel execution.}
}
\description{
Computes K-fold Cross-Validation Accuracy for Twin K svm
}
\author{
Zhang Jiaqi
}
