% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twinsvm_ovr_cv.R
\name{cv.twinsvm_ovr}
\alias{cv.twinsvm_ovr}
\title{An R implementation of Twin-SVM}
\usage{
cv.twinsvm_ovr(
  X,
  y,
  K = 5,
  kernel = c("None", "linear", "rbf", "poly"),
  reg = 1e-07,
  gamma = 1/ncol(X),
  shuffer = TRUE,
  seed = NULL
)
}
\arguments{
\item{X, y}{dataset and label}

\item{K}{`K = 10` means K-fold cross validation}

\item{kernel}{kernel function}

\item{reg}{regularization tern}

\item{gamma}{rbf kernel parameter}

\item{shuffer}{`shuffer = TRUE` means shuffer the dataset}

\item{seed}{is shuffer random seed}
}
\value{
average accuracy in K-fold cross validation
}
\description{
An R implementation of Twin-SVM
}
\examples{
library(manysvms)
data("iris")

X <- iris[, 1:4]
y <- iris[, 5]
cv.twinsvm_ovr(X, y, K = 10, kernel = 'rbf', gamma = 1/8, seed = 1234)
}
\author{
Zhang Jiaqi
}
