% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{f1score}
\alias{f1score}
\title{F1 Score
return F1 Score for two vector}
\usage{
f1score(y, y_hat)
}
\arguments{
\item{y, y_hat}{real values and fitted values.}
}
\description{
F1 Score
return F1 Score for two vector
}
\author{
Zhang Jiaqi
}
