% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{generate_character_label}
\alias{generate_character_label}
\title{Transform Class to Character Class label}
\usage{
generate_character_label(y)
}
\arguments{
\item{y}{label.}
}
\description{
Transform Class to Character Class label
}
\author{
Zhang Jiaqi
}
