% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{glass}
\alias{glass}
\title{Glass Identification Data Set (UCI dataset)}
\format{
A data frame with 214 rows and 10 variables:
\describe{
  \item{RI}{refractive index}
  \item{Na}{Sodium (unit measurement: weight percent in corresponding
  oxide, as are attributes 4-10)}
  \item{Mg}{Magnesium}
  \item{Al}{Aluminum}
  \item{Si}{Silicon}
  \item{K}{Potassium}
  \item{Ca}{Calcium}
  \item{Ba}{Barium}
  \item{Fe}{Iron}
  \item{glass.type}{Type of glass}

}
}
\usage{
glass
}
\description{
From USA Forensic Science Service; 6 types of glass;
defined in terms of their oxide content (i.e. Na, Fe, K, etc)
}
\keyword{datasets}
