% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection.R
\name{grid_search_cv_Xynoisy}
\alias{grid_search_cv_Xynoisy}
\title{Grid Search and Cross Validation with Noisy (Simulation Only)}
\usage{
grid_search_cv_Xynoisy(
  model,
  X,
  y,
  X_noisy,
  y_noisy,
  folds = 5,
  metrics,
  param_list,
  predict_func = predict,
  pipeline = NULL,
  metrics_params = NULL,
  predict_params = NULL,
  model_settings = NULL,
  transy = FALSE,
  shuffle = TRUE,
  seed = NULL,
  model_seed = NULL,
  threads.num = parallel::detectCores() - 1
)
}
\arguments{
\item{model}{your model.}

\item{X, y}{dataset and label.}

\item{X_noisy}{dataset with noise.}

\item{y_noisy}{label (contains label noise)}

\item{folds}{a positive integer indicating the number of folds (sequential
split, compatible with the old \code{K} argument) or a list of
index vectors, where each element contains the test-set row
indices of one fold.}

\item{metrics}{this parameter receive a metric function.}

\item{param_list}{parameter list.}

\item{predict_func}{this parameter receive a function for predict.}

\item{pipeline}{preprocessing pipline.}

\item{metrics_params}{set parameter for each metrics (need a list).}

\item{predict_params}{set parameters for each predict method (need a list).}

\item{model_settings}{set parameters for model (need a list).}

\item{transy}{apply transforms defined in `pipeline` on y, default FALSE.}

\item{shuffle}{if set \code{shuffle==TRUE}, This function will shuffle
the dataset (only used when \code{folds} is a number).}

\item{seed}{random seed for \code{shuffle} option (only used when
\code{folds} is a number).}

\item{model_seed}{random_seed for model.}

\item{threads.num}{the number of threads used for parallel execution.}
}
\value{
return a metric matrix
}
\description{
Grid Search and Cross Validation with Noisy (Simulation Only)
}
\author{
Zhang Jiaqi.
}
