% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_functions.R
\name{kernel_function}
\alias{kernel_function}
\title{Kernel Function}
\usage{
kernel_function(
  x1,
  x2,
  kernel.type = c("linear", "rbf", "poly"),
  gamma = 1/ncol(x1),
  degree = 3,
  coef0 = 0,
  symmetric = FALSE,
  ...
)
}
\arguments{
\item{x1, x2}{input matrices.}

\item{kernel.type}{choose kernel type.}

\item{gamma}{parameter for \code{'rbf'} and \code{'poly'} kernel. Default \code{gamma = 1/ncol(X)}.}

\item{degree}{parameter for polynomial kernel, default: \code{degree = 3}.}

\item{coef0}{parameter for polynomial kernel,  default: \code{coef0 = 0}.}

\item{symmetric}{if \code{x1 == x2} you can set \code{symmetric == TRUE}.}

\item{...}{redundant parameters.}
}
\description{
Kernel Function
}
\author{
Zhang Jiaqi.
}
