% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_systems_solver.R
\name{LLtsolve}
\alias{LLtsolve}
\title{Solve Linear Systems.}
\usage{
LLtsolve(Lt, b)
}
\arguments{
\item{Lt}{a upper triangula square matrix containing the coefficients of the linear system.}

\item{b}{a vector or matrix giving the right-hand side(s).}
}
\description{
\code{LLtsolve} solves the equation {\eqn{L L' x = b}}, where \eqn{L'} is a upper
triangula square matrix
}
\author{
Zhang Jiaqi.
}
