% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lssvr.R
\name{lssvr}
\alias{lssvr}
\title{Least Squares  Support Vector Regression}
\usage{
lssvr(
  X,
  y,
  C = 1,
  kernel = c("linear", "rbf", "poly"),
  gamma = 1/ncol(X),
  degree = 3,
  coef0 = 0,
  rcpp = TRUE
)
}
\arguments{
\item{X, y}{dataset and response variable.}

\item{C}{plenty term.}

\item{kernel}{kernel function.}

\item{gamma}{parameter for \code{'rbf'} and \code{'poly'} kernel. Default \code{gamma = 1/ncol(X)}.}

\item{degree}{parameter for polynomial kernel, default: \code{degree = 3}.}

\item{coef0}{parameter for polynomial kernel,  default: \code{coef0 = 0}.}

\item{rcpp}{speed up your code with Rcpp, default \code{rcpp = TRUE}.}
}
\value{
return lssvr object.
}
\description{
\code{lssvr} is an R implementation of LSSVR
}
\author{
Zhang Jiaqi.
}
