% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{mean_absolute_error}
\alias{mean_absolute_error}
\title{Mean Absolute Error}
\usage{
mean_absolute_error(y, y_hat)
}
\arguments{
\item{y, y_hat}{real values and fitted values.}
}
\description{
Mean Absolute Error
}
\author{
Zhang Jiaqi
}
