% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{mean_squared_error}
\alias{mean_squared_error}
\title{Mean Squared Error
return mean squared error for two vector}
\usage{
mean_squared_error(y, y_hat, dof = 0)
}
\arguments{
\item{y, y_hat}{real values and fitted values.}

\item{dof}{degree of freedom.}
}
\description{
Mean Squared Error
return mean squared error for two vector
}
\author{
Zhang Jiaqi
}
