% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizers.R
\name{nesterov}
\alias{nesterov}
\title{Nesterov's Accelerated Gradient Descent}
\usage{
nesterov(
  X,
  y,
  w,
  m,
  max.steps,
  fx,
  eps = 1e-05,
  v = matrix(0, nrow(w)),
  eta = 1,
  gam = 0.5,
  k = 0.5,
  ...
)
}
\arguments{
\item{X, y}{dataset and label.}

\item{w}{initial point.}

\item{m}{mini-batch size for pegasos solver.}

\item{max.steps}{the number of iterations to solve the optimization problem.}

\item{fx}{sub-gradient of objective function.}

\item{eps}{the precision of the optimization algorithm.}

\item{v}{initial velocity.}

\item{eta}{initial learning rate.}

\item{gam}{momentum parameter.}

\item{k}{decay rate of learning rate.}

\item{...}{additional settings for the sub-gradient.}
}
\value{
return optimal solution.
}
\description{
Nesterov's Accelerated Gradient Descent
}
\author{
Zhang Jiaqi.
}
