% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{noisy_label_generator}
\alias{noisy_label_generator}
\title{Generate noisy label}
\usage{
noisy_label_generator(y, p, seed = NULL)
}
\arguments{
\item{y}{clean label.}

\item{p}{ratio of noisy.}

\item{seed}{random seed.}
}
\description{
Generate noisy label
}
\author{
Zhang Jiaqi
}
