% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_to_multi_classifier.R
\name{OVR_Classifier}
\alias{OVR_Classifier}
\title{Convert Binary Classifier to multiclass classifier
by Using Ones versus Rest Strategy}
\usage{
OVR_Classifier(X, y, model, ...)
}
\arguments{
\item{X, y}{dataset and label.}

\item{model}{binary classifier}

\item{...}{parameter of model.}
}
\value{
return \code{Classifiers_OVR} object.
}
\description{
\code{pegasos} is an R implementation of Pegasos algorithm
}
\author{
Zhang Jiaqi.
}
