% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pin_mbsvm.R
\name{pin_mbsvm}
\alias{pin_mbsvm}
\title{Pinball Loss Multiple Birth Support Vector Machine}
\usage{
pin_mbsvm(
  X,
  y,
  C = 1,
  kernel = c("linear", "rbf", "poly"),
  gamma = 1/ncol(X),
  degree = 3,
  coef0 = 0,
  tau = 1,
  eps = 1e-05,
  max.steps = 4000,
  solver = c("dual"),
  fit_intercept = TRUE,
  randx = 1,
  ...
)
}
\arguments{
\item{X, y}{dataset and label.}

\item{C}{plenty term.}

\item{kernel}{kernel function. The definitions of various kernel functions are as follows:
\describe{
    \item{linear:}{\eqn{u'v}{u'*v}}
    \item{poly:}{\eqn{(\gamma u'v + coef0)^{degree}}{(gamma*u'*v + coef0)^degree}}
    \item{rbf:}{\eqn{e^{(-\gamma |u-v|^2)}}{exp(-gamma*|u-v|^2)}}
}}

\item{gamma}{parameter for \code{'rbf'} and \code{'poly'} kernel. Default \code{gamma = 1/ncol(X)}.}

\item{degree}{parameter for polynomial kernel, default: \code{degree = 3}.}

\item{coef0}{parameter for polynomial kernel,  default: \code{coef0 = 0}.}

\item{tau}{parameter for pinball loss.}

\item{eps}{the precision of the optimization algorithm.}

\item{max.steps}{the number of iterations to solve the optimization problem.}

\item{solver}{\code{"dual"} is available.}

\item{fit_intercept}{if set \code{fit_intercept = TRUE},
the function will evaluates intercept.}

\item{randx}{parameter for reduce SVM, default \code{randx = 1}.}

\item{...}{unused parameters.}
}
\value{
return \code{MBSVMClassifier} object.
}
\description{
\code{pin_mbsvm} is an R implementation of Pin-MBSVM
}
\author{
Zhang Jiaqi.
}
