% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pin_tksvc.R
\name{pin_tksvc}
\alias{pin_tksvc}
\title{Hinge Twin Multi-Class Support Vector Machine}
\usage{
pin_tksvc(
  X,
  y,
  C1 = 1,
  C2 = 1,
  C3 = C1,
  C4 = C2,
  kernel = c("linear", "rbf", "poly"),
  gamma = 1/ncol(X),
  degree = 3,
  coef0 = 0,
  epsilon = 0.1,
  tau1 = 1,
  tau2 = 1,
  tau3 = tau1,
  tau4 = tau2,
  eps = 1e-05,
  max.steps = 5000,
  solver = c("dual"),
  fit_intercept = TRUE,
  randx = 1
)
}
\arguments{
\item{X, y}{dataset and label.}

\item{C1, C2, C3, C4}{plenty term.}

\item{kernel}{kernel function. The definitions of various kernel functions are as follows:
\describe{
    \item{linear:}{\eqn{u'v}{u'*v}}
    \item{poly:}{\eqn{(\gamma u'v + coef0)^{degree}}{(gamma*u'*v + coef0)^degree}}
    \item{rbf:}{\eqn{e^{(-\gamma |u-v|^2)}}{exp(-gamma*|u-v|^2)}}
}}

\item{gamma}{parameter for \code{'rbf'} and \code{'poly'} kernel. Default \code{gamma = 1/ncol(X)}.}

\item{degree}{parameter for polynomial kernel, default: \code{degree = 3}.}

\item{coef0}{parameter for polynomial kernel,  default: \code{coef0 = 0}.}

\item{epsilon}{rest class parameter.}

\item{tau1, tau2, tau3, tau4}{parameter for pinball loss function.}

\item{eps}{the precision of the optimization algorithm.}

\item{max.steps}{the number of iterations to solve the optimization problem.}

\item{solver}{\code{"dual"} is available.}

\item{fit_intercept}{if set \code{fit_intercept = TRUE},
the function will evaluates intercept.}

\item{randx}{parameter for reduce SVM, default \code{randx = 0.1}.}
}
\value{
return \code{HingeSVMClassifier} object.
}
\description{
\code{hinge_tksvc} is an R implementation of Hinge-TKSVC
}
\author{
Zhang Jiaqi.
}
