% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pin_tsvm.R
\name{pin_tsvm}
\alias{pin_tsvm}
\title{Pinball Twin Support Vector Machine}
\usage{
pin_tsvm(
  X,
  y,
  C1 = 1,
  C2 = C1,
  kernel = c("linear", "rbf", "poly"),
  tau1 = 0.5,
  tau2 = tau1,
  gamma = 1/ncol(X),
  degree = 3,
  coef0 = 0,
  eps = 1e-05,
  max.steps = 4000,
  fit_intercept = TRUE,
  reduce_set = NULL
)
}
\arguments{
\item{X, y}{dataset and label.}

\item{C1, C2}{plenty term.}

\item{kernel}{kernel function. The definitions of various kernel functions are as follows:
\describe{
    \item{linear:}{\eqn{u'v}{u'*v}}
    \item{poly:}{\eqn{(\gamma u'v + coef0)^{degree}}{(gamma*u'*v + coef0)^degree}}
    \item{rbf:}{\eqn{e^{(-\gamma |u-v|^2)}}{exp(-gamma*|u-v|^2)}}
}}

\item{tau1, tau2}{parameter for pinball loss function, should be greater than 0.
\code{tau1} is used for the first hyperplane and \code{tau2} for
the second; \code{tau = 0} reduces to the hinge loss.}

\item{gamma}{parameter for \code{'rbf'} and \code{'poly'} kernel. Default \code{gamma = 1/ncol(X)}.}

\item{degree}{parameter for polynomial kernel, default: \code{degree = 3}.}

\item{coef0}{parameter for polynomial kernel,  default: \code{coef0 = 0}.}

\item{eps}{the precision of the optimization algorithm.}

\item{max.steps}{the number of iterations to solve the optimization problem.}

\item{fit_intercept}{if set \code{fit_intercept = TRUE},
the function will evaluates intercept.}

\item{reduce_set}{reduce set for reduce SVM, default \code{reduce_set = NULL}.}
}
\value{
return \code{TSVMClassifier} object.
}
\description{
\code{pin_tsvm} is an R implementation of Pin-TSVM
}
\author{
Zhang Jiaqi.
}
