% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hinge_eps_tsvr.R
\name{plot.EPSTSVRRegressor}
\alias{plot.EPSTSVRRegressor}
\title{Plot Method for Twin Support Vector Regression}
\usage{
\method{plot}{EPSTSVRRegressor}(x, ...)
}
\arguments{
\item{x}{a fitted object of class inheriting from \code{EPSTSVRRegressor}.}

\item{...}{unused parameter.}
}
\description{
Plot Method for Twin Support Vector Regression
}
\author{
Zhang Jiaqi
}
