% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hinge_eps_tsvr.R
\name{plot.EPSTWSVRRegressor}
\alias{plot.EPSTWSVRRegressor}
\title{Plot Method for Twin Support Vector Regression}
\usage{
\method{plot}{EPSTWSVRRegressor}(x, ...)
}
\arguments{
\item{x}{a fitted object of class inheriting from \code{EPSTWSVRRegressor}.}

\item{...}{unused parameter.}
}
\description{
Plot Method for Twin Support Vector Regression
}
\author{
Zhang Jiaqi
}
