% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbsvm.R
\name{plot.mbsvm}
\alias{plot.mbsvm}
\title{Plot Method for Multiple Birth Support Vector Machines}
\usage{
\method{plot}{mbsvm}(x, xlab = "x1", ylab = "x2", ...)
}
\arguments{
\item{x}{Object of class \code{mbsvm}.}

\item{xlab, ylab}{your x-axis name and y-axis name.}

\item{...}{unused parameter.}
}
\description{
If there are only two variables in the training data set,
then you can call this function to draw the decision boundary
for your \code{mbsvm}.
}
\examples{

library(manysvms)
library(MASS)
set.seed(112)
n <- 300
n_c <- 3
sig <- diag(c(0.02, 0.03))
x1 <- mvrnorm(n/n_c, mu = c(-0.6, 0), Sigma = sig)
x2 <- mvrnorm(n/n_c, mu = c(0.6, 0), Sigma = sig)
x3 <- mvrnorm(n/n_c, mu = c(0, 0.6), Sigma = sig)
X <- rbind(x1, x2, x3)
y <- rep(c(1,2,3), rep(n/n_c, n_c))
mbsvm_model = mbsvm(X, y, reg = 1e-3)
plot(mbsvm_model)
}
\author{
Zhang Jiaqi
}
