% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hinge_nhsvm.R
\name{plot.NHSVMClassifier}
\alias{plot.NHSVMClassifier}
\title{Plot Method for Nonparallel Hyperplane Support Vector Machine}
\usage{
\method{plot}{NHSVMClassifier}(x, ...)
}
\arguments{
\item{x}{a fitted object of class inheriting from \code{NHSVMClassifier}.}

\item{...}{unused parameter.}
}
\description{
Plot Method for Nonparallel Hyperplane Support Vector Machine
}
\author{
Zhang Jiaqi
}
