% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hinge_svm.R
\name{plot.SVMClassifier}
\alias{plot.SVMClassifier}
\title{Plot Method for Support Vector Machine}
\usage{
\method{plot}{SVMClassifier}(x, ...)
}
\arguments{
\item{x}{a fitted object of class inheriting from \code{SVMClassifier}.}

\item{...}{unused parameter.}
}
\description{
Plot Method for Support Vector Machine
}
\author{
Zhang Jiaqi
}
