% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hinge_tsvr.R
\name{plot.TSVMRegressor}
\alias{plot.TSVMRegressor}
\title{Plot Method for Twin Support Vector Regression}
\usage{
\method{plot}{TSVMRegressor}(x, ...)
}
\arguments{
\item{x}{a fitted object of class inheriting from \code{TSVRRegressor}.}

\item{...}{unused parameter.}
}
\description{
Plot Method for Twin Support Vector Regression
}
\author{
Zhang Jiaqi
}
