% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{precision}
\alias{precision}
\title{Precision
return precision for two vector}
\usage{
precision(y, y_hat)
}
\arguments{
\item{y, y_hat}{real values and fitted values.}
}
\description{
Precision
return precision for two vector
}
\author{
Zhang Jiaqi
}
