% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_to_multi_classifier.R
\name{predic.OVR_Classifier}
\alias{predic.OVR_Classifier}
\title{Predict Method of OVR Classifier}
\usage{
predic.OVR_Classifier(object, X, y, predict_func = predict, ...)
}
\arguments{
\item{object}{a fitted object of class inheriting from \code{Classifiers_OVR}.}

\item{X, y}{dataset and label.}

\item{predict_func}{predict function of your model.}

\item{...}{parameters of your model.}
}
\value{
return predict results.
}
\description{
\code{pegasos} is an R implementation of Pegasos algorithm
}
\author{
Zhang Jiaqi.
}
