% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svr.R
\name{predict.eps.svr}
\alias{predict.eps.svr}
\title{Predict Method for epsilon - Support Vector Regression}
\usage{
\method{predict}{eps.svr}(object, X, y = NULL, ...)
}
\arguments{
\item{object}{A fitted object of class inheriting from \code{eps.svr}.}

\item{X}{A new data frame for predicting.}

\item{y}{A label data frame corresponding to X.
will print the prediction results and other information.}

\item{...}{unused parameter.}
}
\description{
Predict Method for epsilon - Support Vector Regression
}
\author{
Zhang Jiaqi
}
