% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hinge_svm.R
\name{predict.hinge_svm}
\alias{predict.hinge_svm}
\title{Predict Method for Hinge Support Vector Machine}
\usage{
\method{predict}{hinge_svm}(object, X, y, ...)
}
\arguments{
\item{object}{a fitted object of class inheriting from \code{HingeSVMClassifier}.}

\item{X}{a new data frame for predicting.}

\item{y}{a label data frame corresponding to X.}

\item{...}{unused parameter.}
}
\description{
Predict Method for Hinge Support Vector Machine
}
\author{
Zhang Jiaqi
}
