% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hinge_svm.R
\name{predict.HingeSVMClassifier}
\alias{predict.HingeSVMClassifier}
\title{Predict Method for Hinge Support Vector Machine}
\usage{
\method{predict}{HingeSVMClassifier}(object, X, ...)
}
\arguments{
\item{object}{a fitted object of class inheriting from \code{HingeSVMClassifier}.}

\item{X}{new data for predicting.}

\item{...}{unused parameter.}
}
\description{
Predict Method for Hinge Support Vector Machine
}
\author{
Zhang Jiaqi
}
