% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lssvm_ovr.R
\name{predict.lssvm_ovr}
\alias{predict.lssvm_ovr}
\title{Predict Method for Least Squares Support Vector Machine}
\usage{
\method{predict}{lssvm_ovr}(object, X, y, ...)
}
\arguments{
\item{object}{A fitted object of class inheriting from \code{lssvm_ovr}.}

\item{X}{A new data frame for predicting.}

\item{y}{A label data frame corresponding to X.}

\item{...}{unused parameter.}
}
\description{
Predict Method for Least Squares Support Vector Machine
}
\author{
Zhang Jiaqi
}
