% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbsvm.R
\name{predict.mbsvm}
\alias{predict.mbsvm}
\title{Predict Method for Multiple Birth Support Vector Machines}
\usage{
\method{predict}{mbsvm}(object, X, y = NULL, show.info = TRUE, ...)
}
\arguments{
\item{object}{A fitted object of class inheriting from \code{mbsvm}.}

\item{X}{A new data frame for predicting.}

\item{y}{A label data frame corresponding to X.}

\item{show.info}{If you set \code{show.info = TRUE}, this function
will print the prediction results and other information.}

\item{...}{unused parameter.}
}
\description{
Predict Method for Multiple Birth Support Vector Machines
}
\author{
Zhang Jiaqi
}
