% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hinge_mbsvm.R
\name{predict.MBSVMClassifier}
\alias{predict.MBSVMClassifier}
\title{Predict Method for Multiple Birth Support Vector Machine}
\usage{
\method{predict}{MBSVMClassifier}(object, X, values = FALSE, ...)
}
\arguments{
\item{object}{a fitted object of class inheriting from \code{SVMClassifier}.}

\item{X}{new data for predicting.}

\item{values}{if set \code{values = TRUE}, this function will return predict
values: f = abs(wx+b).}

\item{...}{unused parameter.}
}
\description{
Predict Method for Multiple Birth Support Vector Machine
}
\author{
Zhang Jiaqi
}
