% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hinge_nhsvm.R
\name{predict.NHSVMClassifier}
\alias{predict.NHSVMClassifier}
\title{Predict Method for Nonparallel Hyperplane Support Vector Machine}
\usage{
\method{predict}{NHSVMClassifier}(object, X, values = FALSE, ...)
}
\arguments{
\item{object}{a fitted object of class inheriting from \code{NHSVMClassifier}.}

\item{X}{new data for predicting.}

\item{values}{if set \code{values = TRUE}, this function will return predict
values: f = abs(wx+b).}

\item{...}{unused parameter.}
}
\description{
Predict Method for Nonparallel Hyperplane Support Vector Machine
}
\author{
Zhang Jiaqi
}
