% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pintwinKsvm.R
\name{predict.pintwinKsvm}
\alias{predict.pintwinKsvm}
\title{Predict Method for Twin K Support Vector Machine with Pin-ball loss}
\usage{
\method{predict}{pintwinKsvm}(object, X, y, ...)
}
\arguments{
\item{object}{Object of class `twinKsvm`.}

\item{X}{A new data frame for predicting.}

\item{y}{A label data frame corresponding to X.}

\item{...}{unused parameter.}
}
\description{
Predict Method for Twin K Support Vector Machine with Pin-ball loss
}
\author{
Zhang Jiaqi
}
