% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svm_ovr.R
\name{predict.svm_ovr}
\alias{predict.svm_ovr}
\title{Predict Method for Support Vector Machine}
\usage{
\method{predict}{svm_ovr}(object, X, y, ...)
}
\arguments{
\item{object}{A fitted object of class inheriting from \code{svm_ovr}.}

\item{X}{A new data frame for predicting.}

\item{y}{A label data frame corresponding to X.}

\item{...}{unused parameter.}
}
\description{
Predict Method for Support Vector Machine
}
\author{
Zhang Jiaqi
}
