% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eps_svr.R
\name{predict.SVMRegressor}
\alias{predict.SVMRegressor}
\title{Predict Method for Support Vector Regression}
\usage{
\method{predict}{SVMRegressor}(object, X, ...)
}
\arguments{
\item{object}{a fitted object of class inheriting from \code{SVMRegressor}.}

\item{X}{new data for predicting.}

\item{...}{unused parameter.}
}
\description{
Predict Method for Support Vector Regression
}
\author{
Zhang Jiaqi
}
