% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hinge_tksvc.R
\name{predict.TKSVCClassifier}
\alias{predict.TKSVCClassifier}
\title{Predict Method for Twin Support Vector Machine}
\usage{
\method{predict}{TKSVCClassifier}(object, X, ...)
}
\arguments{
\item{object}{a fitted object of class inheriting from \code{SVMClassifier}.}

\item{X}{new data for predicting.}

\item{...}{unused parameter.}
}
\description{
Predict Method for Twin Support Vector Machine
}
\author{
Zhang Jiaqi
}
