% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hinge_tsvm.R
\name{predict.TSVMClassifier}
\alias{predict.TSVMClassifier}
\title{Predict Method for Twin Support Vector Machine}
\usage{
\method{predict}{TSVMClassifier}(object, X, values = FALSE, ...)
}
\arguments{
\item{object}{a fitted object of class inheriting from \code{TSVMClassifier}.}

\item{X}{new data for predicting.}

\item{values}{if set \code{values = TRUE}, this function will return predict
values: f = abs(wx+b).}

\item{...}{unused parameter.}
}
\description{
Predict Method for Twin Support Vector Machine
}
\author{
Zhang Jiaqi
}
