% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsvr.R
\name{predict.tsvr}
\alias{predict.tsvr}
\title{Predict Method for Multiple Birth Support Vector Machines}
\usage{
\method{predict}{tsvr}(object, X, y, ...)
}
\arguments{
\item{object}{a fitted object of class inheriting from \code{twinsvr}.}

\item{X}{a new data frame for predicting.}

\item{y}{response variable data frame corresponding to X.}

\item{...}{unused parameter.}
}
\description{
Predict Method for Multiple Birth Support Vector Machines
}
\author{
Zhang Jiaqi
}
