% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hinge_tsvr.R
\name{predict.TSVRRegressor}
\alias{predict.TSVRRegressor}
\title{Predict Method for Twin Support Vector Regression}
\usage{
\method{predict}{TSVRRegressor}(object, X, ...)
}
\arguments{
\item{object}{a fitted object of class inheriting from \code{SVMClassifier}.}

\item{X}{new data for predicting.}

\item{...}{unused parameter.}
}
\description{
Predict Method for Twin Support Vector Regression
}
\author{
Zhang Jiaqi
}
