% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twinKsvm.R
\name{predict.twinKsvm}
\alias{predict.twinKsvm}
\title{Predict Method for Twin K Support Vector Machine}
\usage{
\method{predict}{twinKsvm}(object, X, y, ...)
}
\arguments{
\item{object}{Object of class `twinKsvm`.}

\item{X}{A new data frame for predicting.}

\item{y}{A label data frame corresponding to X.}

\item{...}{unused parameter.}
}
\description{
Predict Method for Twin K Support Vector Machine
}
\author{
Zhang Jiaqi
}
