% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twinsvm.R
\name{predict.twinsvm}
\alias{predict.twinsvm}
\title{Predict Method for Multiple Birth Support Vector Machines}
\usage{
\method{predict}{twinsvm}(twinsvm, X, y)
}
\arguments{
\item{twinsvm}{Object of class `twinsvm`.}

\item{X}{A new data frame for predicting.}

\item{y}{A label data frame corresponding to X.}
}
\description{
This function predicts values based upon a model trained by \code{twinsvm}.
}
\examples{
library(manysvms)
data('iris')

m <- 100
X <- iris[0:m, 1:2]
y <- iris[0:m, 5]
model <- twinsvm(X, y, kernel = 'linear')
pred <-predict(model, X, y)
}
\author{
Zhang Jiaqi
}
