% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twinsvm_ovr.R
\name{predict.twinsvm_ovr}
\alias{predict.twinsvm_ovr}
\title{Predict Method for Twin Support Vector Machines (O v R)}
\usage{
\method{predict}{twinsvm_ovr}(object, X, y, ...)
}
\arguments{
\item{object}{Object of class `twinsvm_ovr`.}

\item{X}{A new data frame for predicting.}

\item{y}{A label data frame corresponding to X.}

\item{...}{unused parameter.}
}
\description{
Predict Method for Twin Support Vector Machines (O v R)
}
\author{
Zhang Jiaqi
}
