% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection.R
\name{print.cv_model}
\alias{print.cv_model}
\title{Print Method for Grid-Search and Cross Validation Results}
\usage{
\method{print}{cv_model}(x, ...)
}
\arguments{
\item{x}{object of class \code{eps.svr}.}

\item{...}{unsed argument.}
}
\description{
Print Method for Grid-Search and Cross Validation Results
}
