% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svr.R
\name{print.eps.svr}
\alias{print.eps.svr}
\title{Print Method for epsilon - Support Vector Regression}
\usage{
\method{print}{eps.svr}(x, ...)
}
\arguments{
\item{x}{object of class \code{eps.svr}.}

\item{...}{unsed argument.}
}
\description{
This function prints information about \code{eps.svr} model.
}
