% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twinsvm.R
\name{print.twinsvm}
\alias{print.twinsvm}
\title{Print Method for Twin Support Vector Machines}
\usage{
\method{print}{twinsvm}(x, ...)
}
\arguments{
\item{x}{object of class \code{twinsvm}.}

\item{...}{unsed argument.}
}
\description{
This function prints information about \code{twinsvm} model.
}
