% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{recall}
\alias{recall}
\title{Recall
return recall for two vector}
\usage{
recall(y, y_hat)
}
\arguments{
\item{y, y_hat}{real values and fitted values.}
}
\description{
Recall
return recall for two vector
}
\author{
Zhang Jiaqi
}
