% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizers.R
\name{rmsprop}
\alias{rmsprop}
\title{Rmsprop Optimizer}
\usage{
rmsprop(
  X,
  y,
  w,
  m,
  max.steps,
  fx,
  eps = 1e-05,
  epsilon = 0.001,
  rho = 0.9,
  delta = 1e-05,
  ...
)
}
\arguments{
\item{X, y}{dataset and label.}

\item{w}{initial point.}

\item{m}{mini-batch size for pegasos solver.}

\item{max.steps}{the number of iterations to solve the optimization problem.}

\item{fx}{sub-gradient of objective function.}

\item{eps}{the precision of the optimization algorithm.}

\item{epsilon}{initial stepsize.}

\item{delta}{avoid division by 0.}

\item{...}{additional settings for the sub-gradient.}

\item{eta}{initial learning rate.}

\item{k}{decay rate of learning rate.}
}
\value{
return optimal solution.
}
\description{
Rmsprop Optimizer
}
\author{
Zhang Jiaqi.
}
