% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{root_mean_squared_error}
\alias{root_mean_squared_error}
\title{Root Mean Squared Error
return Root Mean Squared Error for two vector}
\usage{
root_mean_squared_error(y, y_hat, dof = 0)
}
\arguments{
\item{y, y_hat}{real values and fitted values.}

\item{dof}{degree of freedom.}
}
\description{
Root Mean Squared Error
return Root Mean Squared Error for two vector
}
\author{
Zhang Jiaqi
}
