% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sh_eps_tsvr.R
\name{sh_eps_tsvr}
\alias{sh_eps_tsvr}
\title{Squared Hinge Epsilon Twin Support Vector Regression}
\usage{
sh_eps_tsvr(
  X,
  y,
  C1 = 1,
  C2 = C1,
  C3 = 1,
  C4 = C3,
  epsilon1 = 0.1,
  epsilon2 = epsilon1,
  kernel = c("linear", "rbf", "poly"),
  gamma = 1/ncol(X),
  degree = 3,
  coef0 = 0,
  eps = 1e-07,
  max.steps = 4000,
  fit_intercept = TRUE,
  reduce_set = NULL
)
}
\arguments{
\item{X, y}{dataset and label.}

\item{C1, C2}{weight of loss term.}

\item{C3, C4}{weight of regularization term.}

\item{epsilon1, epsilon2}{parameter for epsilon tube.}

\item{kernel}{kernel function. The definitions of various kernel functions are as follows:
\describe{
    \item{linear:}{\eqn{u'v}{u'*v}}
    \item{poly:}{\eqn{(\gamma u'v + coef0)^{degree}}{(gamma*u'*v + coef0)^degree}}
    \item{rbf:}{\eqn{e^{(-\gamma |u-v|^2)}}{exp(-gamma*|u-v|^2)}}
}}

\item{gamma}{parameter for \code{'rbf'} and \code{'poly'} kernel. Default \code{gamma = 1/ncol(X)}.}

\item{degree}{parameter for polynomial kernel, default: \code{degree = 3}.}

\item{coef0}{parameter for polynomial kernel,  default: \code{coef0 = 0}.}

\item{eps}{the precision of the optimization algorithm.}

\item{max.steps}{the number of iterations to solve the optimization problem.}

\item{fit_intercept}{if set \code{fit_intercept = TRUE},
the function will evaluates intercept.}

\item{reduce_set}{reduce set for reduce SVM, default \code{reduce_set = NULL}.}
}
\value{
return \code{TSVRClassifier} object.
}
\description{
\code{sh_eps_tsvr} is an R implementation of Squared-Hinge-EPS-TSVR
}
\author{
Zhang Jiaqi.
}
