% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigmoid_svm.R
\name{sigmoid_svm}
\alias{sigmoid_svm}
\title{Sigmoid Support Vector Machine}
\usage{
sigmoid_svm(
  X,
  y,
  C = 1,
  kernel = c("linear", "rbf", "poly"),
  gamma = 1/ncol(X),
  degree = 3,
  coef0 = 0,
  epsilon = 0,
  lambda = 1,
  eps = 1e-05,
  eps.cccp = 0.01,
  max.steps = 80,
  cccp.steps = 10,
  batch_size = nrow(X)/10,
  solver = c("dual", "primal"),
  rcpp = TRUE,
  fit_intercept = TRUE,
  optimizer = pegasos,
  randx = 0.1,
  ...
)
}
\arguments{
\item{X, y}{dataset and label.}

\item{C}{plenty term.}

\item{kernel}{kernel function. The definitions of various kernel functions are as follows:
\describe{
    \item{linear:}{\eqn{u'v}{u'*v}}
    \item{poly:}{\eqn{(\gamma u'v + coef0)^{degree}}{(gamma*u'*v + coef0)^degree}}
    \item{rbf:}{\eqn{e^{(-\gamma |u-v|^2)}}{exp(-gamma*|u-v|^2)}}
}}

\item{gamma}{parameter for \code{'rbf'} and \code{'poly'} kernel. Default \code{gamma = 1/ncol(X)}.}

\item{degree}{parameter for polynomial kernel, default: \code{degree = 3}.}

\item{coef0}{parameter for polynomial kernel,  default: \code{coef0 = 0}.}

\item{epsilon}{parameter for sigmoid loss (epsilon-insensitive zone).}

\item{lambda}{parameter for sigmoid loss (loss increase speed).}

\item{eps}{the precision of the optimization algorithm.}

\item{eps.cccp}{the precision of the optimization algorithm.}

\item{max.steps}{the number of iterations to solve the optimization problem.}

\item{cccp.steps}{the number of iterations of CCCP.}

\item{batch_size}{mini-batch size for primal solver.}

\item{solver}{\code{"dual"} and \code{"primal"} are available.}

\item{rcpp}{speed up your code with Rcpp, default \code{rcpp = TRUE}.}

\item{fit_intercept}{if set \code{fit_intercept = TRUE},
the function will evaluates intercept.}

\item{optimizer}{default primal optimizer pegasos.}

\item{randx}{parameter for reduce SVM, default \code{randx = 0.1}.}

\item{...}{unused parameters.}
}
\value{
return \code{HingeSVMClassifier} object.
}
\description{
\code{sigmoid_svm} is an R implementation of Sigmoid-SVM
}
\author{
Zhang Jiaqi.
}
