% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svm_ovr.R
\name{svm_ovr}
\alias{svm_ovr}
\title{SVM for Multi-classification by Using Ones versus Rest Strategy}
\usage{
svm_ovr(
  X,
  y,
  C = 1,
  kernel = c("linear", "rbf", "poly"),
  gamma = 1/ncol(X),
  degree = 3,
  coef0 = 0,
  tol = 1e-05,
  max.steps = 200,
  rcpp = TRUE,
  fit_intercept = TRUE
)
}
\arguments{
\item{X, y}{dataset and label.}

\item{C}{plenty term.}

\item{kernel}{kernel function. The definitions of various kernel functions are as follows:
\describe{
    \item{linear:}{\eqn{u'v}{u'*v}}
    \item{poly:}{\eqn{(\gamma u'v + coef0)^{degree}}{(gamma*u'*v + coef0)^degree}}
    \item{rbf:}{\eqn{e^{(-\gamma |u-v|^2)}}{exp(-gamma*|u-v|^2)}}
}}

\item{gamma}{parameter for \code{'rbf'} and \code{'poly'} kernel. Default \code{gamma = 1/ncol(X)}.}

\item{degree}{parameter for polynomial kernel, default: \code{degree = 3}.}

\item{coef0}{parameter for polynomial kernel,  default: \code{coef0 = 0}.}

\item{tol}{the precision of the optimization algorithm.}

\item{max.steps}{the number of iterations to solve the optimization problem.}

\item{rcpp}{speed up your code with Rcpp, default \code{rcpp = TRUE}.}

\item{fit_intercept}{if set \code{fit_intercept = TRUE},
the function will evaluates intercept.}
}
\value{
return \code{svm_ovr} object.
}
\description{
\code{svm_ovr} is an R implementation of support vector machine
}
\author{
Zhang Jiaqi.
}
