% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{train_test_split}
\alias{train_test_split}
\title{Split Training and Testting data}
\usage{
train_test_split(X, y, test_size = 0.3, shuffle = TRUE, seed = NULL)
}
\arguments{
\item{X, y}{dataset and label.}

\item{test_size}{the size of test dataset in [0, 1].}

\item{shuffle}{shuffle the dataset.}

\item{seed}{random seed option.}
}
\description{
Return training and testting dataset.
}
\author{
Zhang Jiaqi.
}
