% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaler.R
\name{trans}
\alias{trans}
\title{Scaler Transform Function}
\usage{
trans(scaler, X)
}
\arguments{
\item{scaler}{data matrix.}

\item{X}{data matrix.}
}
\description{
Scaler Transform Function
}
\author{
Zhang Jiaqi
}
