% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsvr.R
\name{tsvr}
\alias{tsvr}
\title{Twin Support Vector Regression}
\usage{
tsvr(
  X,
  y,
  C1 = 1,
  C2 = 1,
  kernel = c("linear", "rbf", "poly"),
  reg = 1e-05,
  kernel_rect = 1,
  gamma = 1/ncol(X),
  degree = 3,
  coef0 = 0,
  tol = 1e-05,
  max.steps = 300,
  rcpp = TRUE
)
}
\arguments{
\item{X, y}{dataset and response variable.}

\item{C1, C2}{plenty term.}

\item{kernel}{kernel function.}

\item{reg}{regularization term to take care of problems due to ill-conditioning in dual problem.}

\item{kernel_rect}{set kernel size. \code{0<= kernel_rect <= 1}.}

\item{gamma}{parameter for \code{'rbf'} and \code{'poly'} kernel. Default \code{gamma = 1/ncol(X)}.}

\item{degree}{parameter for polynomial kernel, default: \code{degree = 3}.}

\item{coef0}{parameter for polynomial kernel,  default: \code{coef0 = 0}.}

\item{tol}{the precision of the optimization algorithm.}

\item{max.steps}{the number of iterations to solve the optimization problem.}

\item{rcpp}{speed up your code with Rcpp, default \code{rcpp = TRUE}.}
}
\value{
return tsvr object.
}
\description{
\code{tsvr} is an R implementation of TSVR.
}
\references{
Xinjun Peng. TSVR: An efficient Twin Support Vector Machine for
regression, Neural Networks, Volume 23, Issue 3, 2010, Pages 365-372.
}
\author{
Zhang Jiaqi.
}
