% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twinKsvm.R
\name{twinKsvm}
\alias{twinKsvm}
\title{Twin K SVM for Multi-classification by Using Ones versus Rest Strategy}
\usage{
twinKsvm(
  X,
  y,
  Ck = rep(1, length(unique(y))),
  kernel = c("linear", "rbf", "poly"),
  gamma = 1/ncol(X),
  reg = 1,
  kernel_rect = 1,
  tol = 1e-05,
  max.steps = 300,
  rcpp = TRUE
)
}
\arguments{
\item{X, y}{dataset and label}

\item{Ck}{plenty term vector}

\item{kernel}{kernel function}

\item{gamma}{rbf kernel parameter}

\item{reg}{regularization tern}

\item{kernel_rect}{set kernel size. \code{0<= kernel_rect <= 1}}

\item{tol}{the precision of the optimization algorithm.}

\item{max.steps}{the number of iterations to solve the optimization problem.}

\item{rcpp}{speed up your code with Rcpp, default \code{rcpp = TRUE}.}
}
\value{
return twinKsvm object
}
\description{
Twin K SVM for Multi-classification by Using Ones versus Rest Strategy
}
\author{
Zhang Jiaqi
}
