% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twinsvm.R
\name{twinsvm}
\alias{twinsvm}
\title{Twin Support Vector Machines}
\usage{
twinsvm(
  X,
  y,
  C1 = 1,
  C2 = 1,
  kernel = c("linear", "rbf", "poly"),
  gamma = 1/ncol(X),
  reg = 1
)
}
\arguments{
\item{X, y}{dataset and label.}

\item{C1}{plenty term 1.}

\item{C2}{plenty term 2.}

\item{kernel}{kernel function. The definitions of various kernel functions are as follows:
\describe{
    \item{linear:}{\eqn{u'v}{u'*v}}
    \item{poly:}{\eqn{(\gamma u'v + coef0)^{degree}}{(gamma*u'*v + coef0)^degree}}
    \item{rbf:}{\eqn{e^{(-\gamma |u-v|^2)}}{exp(-gamma*|u-v|^2)}}
}}

\item{gamma}{parameter for \code{'rbf'} and \code{'poly'} kernel. Default \code{gamma = 1/ncol(X)}.}

\item{reg}{regularization term to take care of problems due to ill-conditioning in dual problem.}
}
\value{
return twinsvm object.
}
\description{
\code{twinsvm} is an R implementation of Twin SVM.
}
\examples{
library(manysvms)
data('iris')

m <- 100
X <- iris[0:m, 1:2]
y <- iris[0:m, 5]
model <- twinsvm(X, y, kernel = 'linear')

coef(model)
pred <-predict(model, X, y)
}
\author{
Zhang Jiaqi.
}
