% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twinsvm_ovr.R
\name{twinsvm_ovr}
\alias{twinsvm_ovr}
\title{Twin SVM for Multi-classification by Using Ones versus Rest Strategy}
\usage{
twinsvm_ovr(
  X,
  y,
  Ck = rep(1, length(unique(y))),
  kernel = c("linear", "rbf", "poly"),
  gamma = 1/ncol(X),
  reg = 1
)
}
\arguments{
\item{X, y}{dataset and label}

\item{Ck}{plenty term vector}

\item{kernel}{kernel function}

\item{gamma}{rbf kernel parameter}

\item{reg}{regularization tern}
}
\value{
return twinsvm object
}
\description{
Twin SVM for Multi-classification by Using Ones versus Rest Strategy
}
\examples{

library(manysvms)
data('iris')

m <- 100
X <- iris[0:m, 1:2]
y <- iris[0:m, 5]
model <- twinsvm_ovr(X, y, kernel = 'linear')

coef(model)
pred <-predict(model, X, y)
}
\author{
Zhang Jiaqi
}
