% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wave_svm.R
\name{wave_svm}
\alias{wave_svm}
\title{Support Vector Machine With Wave Loss Function}
\usage{
wave_svm(
  X,
  y,
  C = 1,
  kernel = c("linear", "rbf", "poly"),
  gamma = 1/ncol(X),
  degree = 3,
  coef0 = 0,
  a = 1,
  lambda = 1,
  max.steps = 4000,
  batch_size = nrow(X)/10,
  fit_intercept = TRUE,
  optimizer = adam,
  reduce_set = NULL,
  ...
)
}
\arguments{
\item{X, y}{dataset and label.}

\item{C}{plenty term.}

\item{kernel}{kernel function. The definitions of various kernel functions are as follows:
\describe{
    \item{linear:}{\eqn{u'v}{u'*v}}
    \item{poly:}{\eqn{(\gamma u'v + coef0)^{degree}}{(gamma*u'*v + coef0)^degree}}
    \item{rbf:}{\eqn{e^{(-\gamma |u-v|^2)}}{exp(-gamma*|u-v|^2)}}
}}

\item{gamma}{parameter for \code{'rbf'} and \code{'poly'} kernel. Default \code{gamma = 1/ncol(X)}.}

\item{degree}{parameter for polynomial kernel, default: \code{degree = 3}.}

\item{coef0}{parameter for polynomial kernel,  default: \code{coef0 = 0}.}

\item{a, lambda}{parameter for wave loss.}

\item{max.steps}{the number of iterations to solve the optimization problem.}

\item{batch_size}{mini-batch size for primal solver.}

\item{fit_intercept}{if set \code{fit_intercept = TRUE},
the function will evaluates intercept.}

\item{optimizer}{default primal optimizer pegasos.}

\item{reduce_set}{reduce set for reduce SVM, default \code{reduce_set = NULL}.}

\item{...}{unused parameters.}
}
\value{
return \code{SVMClassifier} object.
}
\description{
\code{wave_svm} is an R implementation of Wave-SVM
}
\author{
Zhang Jiaqi.
}
