% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection.R
\name{cross_validation}
\alias{cross_validation}
\title{K-Fold Cross Validation}
\usage{
cross_validation(model, X, y, K = 5, metric, predict_func = predict, ...)
}
\arguments{
\item{model}{your model.}

\item{X, y}{dataset and label.}

\item{K}{number of folds.}

\item{metric}{this parameter receive a metric function.}

\item{predict_func}{this parameter receive a function for predict.}

\item{...}{additional parameters for your model.}

\item{shuffle}{if set \code{shuffle==TRUE}, This function will shuffle the dataset.}

\item{seed}{random seed for \code{shuffle} option.}
}
\value{
return a metric matrix
}
\description{
K-Fold Cross Validation
}
\author{
Zhang Jiaqi.
}
