% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection.R
\name{cross_validation}
\alias{cross_validation}
\title{K-Fold Cross Validation}
\usage{
cross_validation(
  model,
  X,
  y,
  K = 5,
  metrics,
  predict_func = predict,
  pipeline = NULL,
  metrics_params = NULL,
  predict_params = NULL,
  model_settings = NULL
)
}
\arguments{
\item{model}{your model.}

\item{X, y}{dataset and label.}

\item{K}{number of folds.}

\item{metrics}{this parameter receive a metric function.}

\item{predict_func}{this parameter receive a function for predict.}

\item{pipeline}{preprocessing pipline.}

\item{metrics_params}{set parameters for each metrics (need a list).}

\item{predict_params}{set parameters for each predict method (need a list).}

\item{model_settings}{set parameters for model (need a list).}
}
\value{
return a metric matrix
}
\description{
K-Fold Cross Validation
}
\author{
Zhang Jiaqi.
}
