% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cskb_svm.R
\name{cskb_svm}
\alias{cskb_svm}
\title{Hinge Support Vector Machine}
\usage{
cskb_svm(
  X,
  y,
  a = 1,
  b = 1,
  C = 1,
  kernel = c("linear", "rbf", "poly"),
  gamma = 1/ncol(X),
  degree = 3,
  coef0 = 0,
  eps = 0.01,
  max.steps = 80,
  batch_size = nrow(X)/10,
  fit_intercept = TRUE,
  randx = 0.1,
  ...
)
}
\arguments{
\item{X, y}{dataset and label.}

\item{a, b, c}{parameters for blinex loss.}

\item{C}{plenty term.}

\item{kernel}{kernel function. The definitions of various kernel functions are as follows:
\describe{
    \item{linear:}{\eqn{u'v}{u'*v}}
    \item{poly:}{\eqn{(\gamma u'v + coef0)^{degree}}{(gamma*u'*v + coef0)^degree}}
    \item{rbf:}{\eqn{e^{(-\gamma |u-v|^2)}}{exp(-gamma*|u-v|^2)}}
}}

\item{gamma}{parameter for \code{'rbf'} and \code{'poly'} kernel. Default \code{gamma = 1/ncol(X)}.}

\item{degree}{parameter for polynomial kernel, default: \code{degree = 3}.}

\item{coef0}{parameter for polynomial kernel,  default: \code{coef0 = 0}.}

\item{eps}{the precision of the optimization algorithm.}

\item{max.steps}{the number of iterations to solve the optimization problem.}

\item{batch_size}{mini-batch size for primal solver.}

\item{fit_intercept}{if set \code{fit_intercept = TRUE},
the function will evaluates intercept.}

\item{randx}{parameter for reduce SVM, default \code{randx = 0.1}.}

\item{...}{unused parameters.}
}
\value{
return \code{HingeSVMClassifier} object.
}
\description{
\code{hinge_svm} is an R implementation of Hinge-SVM
}
\author{
Li Feihong
}
