% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbsvm.R
\name{cv.mbsvm}
\alias{cv.mbsvm}
\title{Computes K-fold Cross-Validated Accuracy for MBSVM}
\usage{
cv.mbsvm(
  X,
  y,
  K = 5,
  kernel = c("linear", "rbf", "poly"),
  reg = 1,
  gamma = 1/ncol(X),
  kernel_rect = 1,
  shuffer = TRUE,
  seed = NULL
)
}
\arguments{
\item{X}{A new data frame for predicting.}

\item{y}{A label data frame corresponding to X.}

\item{K}{Number of folds.}

\item{kernel}{kernel type. Default value \code{kernel = 'linear'}.}

\item{reg}{regularization term to take care of problems due to ill-conditioning in dual problem.}

\item{gamma}{parameter needed for rbf kernel.}

\item{kernel_rect}{set kernel size. \code{0<= kernel_rect <= 1}}

\item{shuffer}{if set \code{shuffer==TRUE}, This function will shuffle the dataset.}

\item{seed}{random seed for \code{shuffer} option.}
}
\description{
Computes K-fold Cross-Validated Accuracy for MBSVM
}
\author{
Zhang Jiaqi
}
