% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ramptwinKsvm.R
\name{cv.ramptwinKsvm}
\alias{cv.ramptwinKsvm}
\title{Computes K-fold Cross-Validation Accuracy for ramp Twin K svm}
\usage{
cv.ramptwinKsvm(
  X,
  y,
  K = 5,
  C = 1,
  s = 0.5,
  kernel = c("linear", "rbf", "poly"),
  gamma = 1/ncol(X),
  degree = 3,
  coef0 = 0,
  reg = 1e-07,
  kernel_rect = 1,
  eps = 0.1,
  tol = 1e-05,
  max.steps = 200,
  step_cccp = 30,
  rcpp = TRUE,
  shuffle = TRUE,
  seed = NULL,
  threads.num = parallel::detectCores() - 1
)
}
\arguments{
\item{X, y}{dataset and label.}

\item{K}{number of folds.}

\item{C}{plenty term vector.}

\item{s}{parameter for ramp loss.}

\item{kernel}{kernel function.}

\item{gamma}{rbf kernel parameter.}

\item{degree}{parameter for polynomial kernel, default: \code{degree = 3}.}

\item{coef0}{parameter for polynomial kernel,  default: \code{coef0 = 0}.}

\item{reg}{regularization term.}

\item{kernel_rect}{set kernel size. \code{0<= kernel_rect <= 1}}

\item{eps}{parameter for rest class.}

\item{tol}{the precision of the optimization algorithm.}

\item{max.steps}{the number of iterations to solve the optimization problem.}

\item{step_cccp}{the number of iterations of Concave–Convex Procedure (CCCP).}

\item{rcpp}{speed up your code with Rcpp, default \code{rcpp = TRUE}.}

\item{shuffle}{if set \code{shuffer==TRUE}, This function will shuffle the dataset.}

\item{seed}{random seed for \code{shuffer} option.}

\item{threads.num}{The number of threads used for parallel execution.}
}
\description{
Computes K-fold Cross-Validation Accuracy for ramp Twin K svm
}
\author{
Zhang Jiaqi
}
