% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twinsvm_ovr_cv.R
\name{cv.twinsvm_ovr}
\alias{cv.twinsvm_ovr}
\title{Twin SVM for Multi-classification by Using Ones versus Rest Strategy
An R implementation of \code{twinsvm} (Cross Validation)}
\usage{
cv.twinsvm_ovr(
  X,
  y,
  K = 5,
  C = 1,
  kernel = c("linear", "rbf", "poly"),
  gamma = 1/ncol(X),
  degree = 3,
  coef0 = 0,
  reg = 1e-07,
  kernel_rect = 1,
  shuffle = TRUE,
  seed = NULL,
  tol = 1e-05,
  max.steps = 200,
  rcpp = TRUE,
  threads.num = parallel::detectCores() - 1
)
}
\arguments{
\item{X, y}{dataset and label}

\item{K}{`K = 10` means K-fold cross validation}

\item{C}{plenty term.}

\item{kernel}{kernel function}

\item{gamma}{parameter for \code{'rbf'} and \code{'poly'} kernel. Default \code{gamma = 1/ncol(X)}.}

\item{degree}{parameter for polynomial kernel, default: \code{degree = 3}.}

\item{coef0}{parameter for polynomial kernel,  default: \code{coef0 = 0}.}

\item{reg}{regularization term to take care of problems due to ill-conditioning in dual problem.}

\item{kernel_rect}{set kernel size. \code{0<= kernel_rect <= 1}}

\item{shuffle}{if set \code{shuffer==TRUE}, This function will shuffle the dataset.}

\item{seed}{random seed for \code{shuffle} option.}

\item{tol}{the precision of the optimization algorithm.}

\item{max.steps}{the number of iterations to solve the optimization problem.}

\item{rcpp}{speed up your code with Rcpp, default \code{rcpp = TRUE}.}

\item{threads.num}{The number of threads used for parallel execution.}
}
\description{
Twin SVM for Multi-classification by Using Ones versus Rest Strategy
An R implementation of \code{twinsvm} (Cross Validation)
}
\examples{
library(manysvms)
data("iris")

X <- iris[, 1:4]
y <- iris[, 5]
cv.twinsvm_ovr(X, y, K = 10, kernel = 'rbf',
               gamma = 1/8, seed = 1234, threads.num = 2)
}
\author{
Zhang Jiaqi
}
