% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{geometric_mean_of_recall}
\alias{geometric_mean_of_recall}
\title{Geometric Mean of Recall}
\usage{
geometric_mean_of_recall(y, y_hat, sample_weight = NULL, correction = 0)
}
\arguments{
\item{y, y_hat}{real values and fitted values.}

\item{sample_weight}{numeric vector of sample weights, default \code{NULL}.}

\item{correction}{weight applied to classes with no true samples, default 0.}
}
\value{
geometric mean of per-class recall, a numeric scalar.
}
\description{
Compute the geometric mean of per-class recall. This implementation aligns
with \code{sklearn.metrics.geometric_mean_score}.
}
\author{
Zhang Jiaqi
}
