% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection.R
\name{grid_search_cv}
\alias{grid_search_cv}
\title{Grid Search and Cross Validation}
\usage{
grid_search_cv(
  model,
  X,
  y,
  K = 5,
  metric,
  param_list,
  predict_func = predict,
  shuffle = TRUE,
  seed = NULL,
  threads.num = parallel::detectCores() - 1,
  ...
)
}
\arguments{
\item{model}{your model.}

\item{X, y}{dataset and label.}

\item{K}{number of folds.}

\item{metric}{this parameter receive a metric function.}

\item{param_list}{parameter list.}

\item{predict_func}{this parameter receive a function for predict.}

\item{shuffle}{if set \code{shuffle==TRUE}, This function will shuffle
the dataset.}

\item{seed}{random seed for \code{shuffle} option.}

\item{threads.num}{the number of threads used for parallel execution.}

\item{...}{additional parameters for your model.}
}
\value{
return a metric matrix
}
\description{
Grid Search and Cross Validation
}
\author{
Zhang Jiaqi.
}
