% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection.R
\name{grid_search_cv_noisy}
\alias{grid_search_cv_noisy}
\title{Grid Search and Cross Validation with Noisy (Simulation Only)}
\usage{
grid_search_cv_noisy(
  model,
  X,
  y,
  y_noisy,
  K = 5,
  metrics,
  param_list,
  predict_func = predict,
  pipeline = NULL,
  metrics_params = NULL,
  predict_params = NULL,
  model_settings = NULL,
  transy = FALSE,
  shuffle = TRUE,
  seed = NULL,
  threads.num = parallel::detectCores() - 1
)
}
\arguments{
\item{model}{your model.}

\item{X, y}{dataset and label.}

\item{y_noisy}{label (contains label noise)}

\item{K}{number of folds.}

\item{metrics}{this parameter receive a metric function.}

\item{param_list}{parameter list.}

\item{predict_func}{this parameter receive a function for predict.}

\item{pipeline}{preprocessing pipline.}

\item{metrics_params}{set parameter for each metrics (need a list).}

\item{predict_params}{set parameters for each predict method (need a list).}

\item{model_settings}{set parameters for model (need a list).}

\item{transy}{apply transforms defined in `pipeline` on y, default FALSE.}

\item{shuffle}{if set \code{shuffle==TRUE}, This function will shuffle
the dataset.}

\item{seed}{random seed for \code{shuffle} option.}

\item{threads.num}{the number of threads used for parallel execution.}
}
\value{
return a metric matrix
}
\description{
Grid Search and Cross Validation with Noisy (Simulation Only)
}
\author{
Zhang Jiaqi.
}
