% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbsvm.R
\name{mbsvm}
\alias{mbsvm}
\title{Multiple Birth SVM}
\usage{
mbsvm(
  X,
  y,
  Ck = rep(1, length(unique(y))),
  kernel = c("linear", "rbf", "poly"),
  gamma = 1/ncol(X),
  reg = 1,
  kernel_rect = 1
)
}
\arguments{
\item{X, y}{dataset and label.}

\item{Ck}{plenty term list.}

\item{kernel}{kernel function.}

\item{gamma}{parameter for \code{'rbf'} and \code{'poly'} kernel. Default \code{gamma = 1/ncol(X)}.}

\item{reg}{regularization term to take care of problems due to ill-conditioning in dual problem.}

\item{kernel_rect}{set kernel size. \code{0<= kernel_rect <= 1}}
}
\value{
return mbsvm object.
}
\description{
\code{mbsvm} is an R implementation of multiple birth SVM
}
\examples{
library(manysvms)
data('iris')


X <- iris[1:100, 1:2]
y <- iris[1:100, 5]

model <- mbsvm(X, y, kernel = 'linear')
pred <-predict(model, X, y)
}
\author{
Zhang Jiaqi.
}
