% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbsvm.R
\name{mbsvm}
\alias{mbsvm}
\title{Multiple Birth Support Vector Machine}
\usage{
mbsvm(
  X,
  y,
  Ck = rep(1, length(unique(y))),
  kernel = c("linear", "rbf", "poly"),
  gamma = 1/ncol(X),
  degree = 3,
  coef0 = 0,
  reg = 1e-07,
  kernel_rect = 1,
  tol = 1e-05,
  max.steps = 200,
  rcpp = TRUE
)
}
\arguments{
\item{X, y}{dataset and label.}

\item{Ck}{plenty term list.}

\item{kernel}{kernel function. The definitions of various kernel functions are as follows:
\describe{
    \item{linear:}{\eqn{u'v}{u'*v}}
    \item{poly:}{\eqn{(\gamma u'v + coef0)^{degree}}{(gamma*u'*v + coef0)^degree}}
    \item{rbf:}{\eqn{e^{(-\gamma |u-v|^2)}}{exp(-gamma*|u-v|^2)}}
}}

\item{gamma}{parameter for \code{'rbf'} and \code{'poly'} kernel. Default \code{gamma = 1/ncol(X)}.}

\item{degree}{parameter for polynomial kernel, default: \code{degree = 3}.}

\item{coef0}{parameter for polynomial kernel,  default: \code{coef0 = 0}.}

\item{reg}{regularization term to take care of problems due to ill-conditioning in dual problem.}

\item{kernel_rect}{set kernel size. \code{0<= kernel_rect <= 1}}

\item{tol}{the precision of the optimization algorithm.}

\item{max.steps}{the number of iterations to solve the optimization problem.}

\item{rcpp}{speed up your code with Rcpp, default \code{rcpp = TRUE}.}
}
\value{
return mbsvm object.
}
\description{
\code{mbsvm} is an R implementation of multiple birth SVM
}
\examples{
library(manysvms)
data('iris')


X <- iris[1:150, 1:4]
y <- iris[1:150, 5]

model1 <- mbsvm(X, y, kernel = 'linear')
pred <-predict(model1, X, y)
}
\author{
Zhang Jiaqi.
}
