% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizers.R
\name{nesterov}
\alias{nesterov}
\title{Nesterov's Accelerated Gradient Descent}
\usage{
nesterov(
  X,
  y,
  w,
  m,
  max.steps,
  fx,
  pars,
  v = matrix(0, nrow(w)),
  lr = 0.01,
  gam = 0.9,
  decay_option = exp_decay,
  ...
)
}
\arguments{
\item{X, y}{dataset and label.}

\item{w}{initial point.}

\item{m}{mini-batch size for pegasos solver.}

\item{max.steps}{the number of iterations to solve the optimization problem.}

\item{fx}{sub-gradient of objective function.}

\item{pars}{parameters list for the sub-gradient.}

\item{v}{initial velocity.}

\item{lr}{initial learning rate.}

\item{gam}{momentum parameter.}

\item{decay_option}{decay option for learning rate.}

\item{...}{additional settings for the sub-gradient.}
}
\value{
return optimal solution.
}
\description{
Nesterov's Accelerated Gradient Descent
}
\author{
Zhang Jiaqi.
}
