% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizers.R
\name{pegasos}
\alias{pegasos}
\title{Primal Estimated sub-Gradient SOlver for SVM (Pegasos)}
\usage{
pegasos(X, y, w, m, max.steps, fx, pars, projection = TRUE, ...)
}
\arguments{
\item{X, y}{dataset and label.}

\item{w}{initial point.}

\item{m}{mini-batch size for pegasos solver.}

\item{max.steps}{the number of iterations to solve the optimization problem.}

\item{fx}{sub-gradient of objective function.}

\item{pars}{parameters list for the sub-gradient.}

\item{projection}{projection option.}

\item{...}{additional settings for the sub-gradient.}
}
\value{
return optimal solution.
}
\description{
Primal Estimated sub-Gradient SOlver for SVM (Pegasos)
}
\references{
${1:Pegasos: Primal Estimated sub-GrAdient SOlver for SVM}
}
\author{
Zhang Jiaqi.
}
