% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hinge_svm.R
\name{predict.SVMClassifier}
\alias{predict.SVMClassifier}
\title{Predict Method for Support Vector Machine}
\usage{
\method{predict}{SVMClassifier}(object, X, values = FALSE, ...)
}
\arguments{
\item{object}{a fitted object of class inheriting from \code{SVMClassifier}.}

\item{X}{new data for predicting.}

\item{values}{if set \code{values = TRUE}, this function will return predict
values: f = wx+b.}

\item{...}{unused parameter.}
}
\description{
Predict Method for Support Vector Machine
}
\author{
Zhang Jiaqi
}
