% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twinKsvm.R
\name{twinKsvm}
\alias{twinKsvm}
\alias{print.twinKsvm}
\title{Twin K Support Vector Machine for Multi-classification}
\usage{
twinKsvm(
  X,
  y,
  Ck = rep(1, 4),
  kernel = c("linear", "rbf", "poly"),
  gamma = 1/ncol(X),
  degree = 3,
  coef0 = 0,
  reg = 1e-07,
  kernel_rect = 1,
  eps = 0.1,
  tol = 1e-05,
  max.steps = 200,
  rcpp = TRUE
)
}
\arguments{
\item{X, y}{dataset and label.}

\item{Ck}{plenty term vector.}

\item{kernel}{kernel function.}

\item{gamma}{rbf kernel parameter.}

\item{degree}{parameter for polynomial kernel, default: \code{degree = 3}.}

\item{coef0}{parameter for polynomial kernel,  default: \code{coef0 = 0}.}

\item{reg}{regularization term.}

\item{kernel_rect}{set kernel size. \code{0<= kernel_rect <= 1}}

\item{eps}{parameter for rest class.}

\item{tol}{the precision of the optimization algorithm.}

\item{max.steps}{the number of iterations to solve the optimization problem.}

\item{rcpp}{speed up your code with Rcpp, default \code{rcpp = TRUE}.}
}
\value{
return twinKsvm object
}
\description{
Twin K Support Vector Machine for Multi-classification
}
\author{
Zhang Jiaqi
}
